+++ frontend/index.html
@@ -8,7 +8,7 @@
   </head>
   <body>
     <div id="root"></div>
-    <script type="module" src="./src/main.tsx"></script>
+    <script type="module" src="./src/app/main.tsx"></script>
     <noscript>
       <div style="font-family: var(--font-text); padding: 4rem">
         <h1
